+++ core/pfe-core/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @patternfly/pfe-core
 
+## 2.4.1
+
+### Patch Changes
+
+- 5b16b3b04: `SlotController`: ensure first render is correct when used in certain javascript frameworks
+
+## 2.4.0
+
+### Minor Changes
+
+- a81bcb133: **Controllers**: Added timestamp controller
+
+### Patch Changes
+
+- 7055add74: `FloatingDOMController`: fixed an incorrect typescript import
+
 ## 2.3.1
 
 ### Patch Changes